% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{emoticons}
\alias{emoticons}
\title{List of 842 Emoticons formatted in UTF-8 with descrpiton}
\format{A data frame with 842 rows and 3 variables:
\describe{
  \item{emoticon}{UTF-8 formatted Emoticon}
  \item{unicode}{Unicode formatted Emoticon}
  \item{bytes.utf.8}{Bytes}
  \item{description}{Description of the Emoticon}
}}
\source{
\url{http://apps.timwhitlock.info/emoji/tables/unicode}
}
\usage{
emoticons
}
\description{
Source Raw Data: http://apps.timwhitlock.info/emoji/tables/unicode
}
\keyword{datasets}

