% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_texts.R
\name{wapp_count_emoticons}
\alias{wapp_count_emoticons}
\title{Count occurence of (unique) emoticons an author has used in an imported chat history}
\usage{
wapp_count_emoticons(chat, author, singular)
}
\arguments{
\item{chat}{a WhatsApp chat history that was imported using \code{\link{wapp_import}}}

\item{author}{name of author in a WhatsApp chat history}

\item{singular}{TRUE for only counting unique emoticons and FALSE for counting all posted emoticons}
}
\value{
Number of (unique) emoticons are used by the choosen 'author'
}
\description{
Count occurence of (unique) emoticons an author has used in an imported chat history
}

