% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{wapp_divide}
\alias{wapp_divide}
\title{Returns a Dataframe that only contains posts with type 'media' oder 'message'}
\usage{
wapp_divide(chat, type)
}
\arguments{
\item{chat}{Chat history that was imported with function wapp_import}

\item{type}{String: 'media' for returning only media; 'message' for returning only pure messages without media}
}
\value{
Dataframe only with chats from the type 'media' or 'message'
}
\description{
Returns a Dataframe that only contains posts with type 'media' oder 'message'
}
\examples{
friends <- wapp_import("Data/WhatsApp Chat mit Freunde Uni.txt", language = "de", techuser = FALSE)
}

