% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{wapp_filter}
\alias{wapp_filter}
\title{Return rows with matching conditions: if type is 'media' or 'message'}
\usage{
wapp_filter(chat, type)
}
\arguments{
\item{chat}{a WhatsApp chat history that was imported using \code{\link{wapp_import}}}

\item{type}{a string. 'message' for returning only pure messages without media. 'media' for returning only media;}
}
\value{
an object of the same class as chat
}
\description{
Return rows with matching conditions: if type is 'media' or 'message'
}

