% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{wapp_prep_author}
\alias{wapp_prep_author}
\title{Replaces author names}
\usage{
wapp_prep_author(author, author.old, author.new)
}
\arguments{
\item{author}{column with names of authors in a WhatsApp chat history that was imported using \code{\link{wapp_import}}}

\item{author.old}{a vector with the 'old' names of authors that should be replaced}

\item{author.new}{a vector with the 'new' names of authors that should be replaced. Positions in vector has to correspond to positions in 'author.old'}
}
\value{
a column with replaced names of authors
}
\description{
Replaces author names
}

