% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{wapp_stat_author}
\alias{wapp_stat_author}
\title{Aggregegates author}
\usage{
wapp_stat_author(chat)
}
\arguments{
\item{chat}{Chat history that was imported with function wapp_import}
}
\value{
Grouped Dataframe with Informations which Author writes how much posts
}
\description{
Aggregegates author
}

