% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{wapp_stat_date}
\alias{wapp_stat_date}
\title{Group an imported chat history by date}
\usage{
wapp_stat_date(chat, n)
}
\arguments{
\item{chat}{a WhatsApp chat history that was imported using \code{\link{wapp_import}}}

\item{n}{number of days}
}
\value{
Grouped dataframe with informations in which time interval how many posts and what is proportion of total posts
}
\description{
Group an imported chat history by date
}

