% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_texts.R
\name{wapp_stat_emoticon}
\alias{wapp_stat_emoticon}
\title{Emoticons per Author}
\usage{
wapp_stat_emoticon(chat, author = "John Doe", n = 3)
}
\arguments{
\item{chat}{Chat history that was imported with function wapp_import}
}
\value{
Grouped Dataframe with Informations which Author writes how much posts
}
\description{
Emoticons per Author
}

