% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_texts.R
\name{wapp_stat_emoticons}
\alias{wapp_stat_emoticons}
\title{Group an imported chat history by emoticons an author has used}
\usage{
wapp_stat_emoticons(chat, author, n = 3)
}
\arguments{
\item{chat}{a WhatsApp chat history that was imported using \code{\link{wapp_import}}}

\item{author}{name of author in a WhatsApp chat history}

\item{n}{number of top 'n' emoticons}
}
\value{
Grouped Dataframe with informations which top 'n' emoticons are used by the choosen 'author'
}
\description{
Group an imported chat history by emoticons an author has used
}

