% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{wapp_stat_firstpost}
\alias{wapp_stat_firstpost}
\title{Who posts the first post on an day}
\usage{
wapp_stat_firstpost(chat)
}
\arguments{
\item{chat}{a WhatsApp chat history that was imported using \code{\link{wapp_import}}}
}
\value{
Grouped dataframe with informations which author has posted how often the first entry of the day
}
\description{
Who posts the first post on an day
}

