% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{wapp_stat_time}
\alias{wapp_stat_time}
\title{Aggregegates Time intervall}
\usage{
wapp_stat_time(chat, intervall)
}
\arguments{
\item{chat}{Chat history that was imported with function wapp_import}

\item{intervall}{String: 'media' for returning only media; 'message' for returning only pure messages without media}
}
\value{
Grouped Dataframe with Informations in which time intervall the most posts occur
}
\description{
Aggregegates Time intervall
}

