% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_texts.R
\name{wapp_stat_word}
\alias{wapp_stat_word}
\title{Word Occurence}
\usage{
wapp_stat_word(chat, word)
}
\arguments{
\item{chat}{Chat history that was imported with function wapp_import}

\item{word}{which word occurence}
}
\value{
Grouped Dataframe with Informations which Author writes how much posts
}
\description{
Word Occurence
}

