% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhatsAppR.R
\name{WhatsAppR}
\alias{WhatsAppR}
\title{Visualising WhatsApp conversation data}
\usage{
WhatsAppR(file, save = FALSE, OS = "android", wordcloud = TRUE)
}
\arguments{
\item{file}{the directory of the WhatsApp .txt file}

\item{save}{a logcal indicating if the output plots should be saved to the current directory}

\item{OS}{indicating if the operating system is "android" or "iOS"}

\item{wordcloud}{a logical indicating if a wordcloud should be plotted or not}
}
\description{
Visualising WhatsApp conversation data
}
\author{
Finlay Campbell <f.campbell15@imperial.ac.uk>
}

