% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wr.data.R
\name{wr.data}
\alias{wr.data}
\title{Reading WhatsApp conversation data}
\usage{
wr.data(file, OS, CheckForErrors = FALSE)
}
\arguments{
\item{file}{the directory of the WhatsApp .txt file}

\item{OS}{indicating if the operating system is "android" or "iOS"}

\item{CheckForErrors}{will check for errors in reading of input data, mostly due to quoted text}
}
\value{
Returns a data frame with formatted WhatsApp data
}
\description{
Reading WhatsApp conversation data
}
\author{
Finlay Campbell <f.campbell15@imperial.ac.uk>
}
