% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{wapp_import}
\alias{wapp_import}
\title{Imports an exported WhatsApp chat history}
\usage{
wapp_import(history, language = "de", techuser = FALSE)
}
\arguments{
\item{history}{File path that contains the chat history.}

\item{language}{Only German Language is supported.}

\item{techuser}{Boolean: If WhatsApp TechUser-Posts should be part of the Dataframe}
}
\value{
Dataframe with the WhatsApp chat history
}
\description{
Imports an exported WhatsApp chat history
}

