% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{wapp_import}
\alias{wapp_import}
\title{Imports an exported WhatsApp chat history}
\usage{
wapp_import(history, os, language = "de", techuser = FALSE)
}
\arguments{
\item{history}{a file path (txt-file) that contains the exported WhatsApp history}

\item{os}{the operating system of the smartphone 'iOS' or 'android'}

\item{language}{Note: Currently only the German language ('de') is supported}

\item{techuser}{a boolean variable. If WhatsApp TechUser-Posts should be part of the dataframe or not}
}
\value{
a dataframe with the imported and processed WhatsApp history
}
\description{
Imports an exported WhatsApp chat history
}

