% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{wapp_prep_author}
\alias{wapp_prep_author}
\title{Replaces author names}
\usage{
wapp_prep_author(author, author.old, author.new)
}
\arguments{
\item{author}{column of author importet}

\item{author.old}{old names}

\item{author.new}{new names}
}
\value{
Adds
}
\description{
Replaces author names
}
\examples{
test
}

