% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{wapp_prep_time}
\alias{wapp_prep_time}
\title{Transformation of Time Values: Adds Column moment.intervall}
\usage{
wapp_prep_time(moment, intervall)
}
\arguments{
\item{intervall}{hour or quartar}

\item{chat}{Chat history that was imported with function wapp_import}
}
\value{
Adds
}
\description{
Transformation of Time Values: Adds Column moment.intervall
}
\examples{
test
}

