% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparation.R
\name{wapp_prep_time}
\alias{wapp_prep_time}
\title{Conversion of time data in 'moment' column to time intervals}
\usage{
wapp_prep_time(moment, interval)
}
\arguments{
\item{moment}{column with time data in a WhatsApp chat history that was imported using \code{\link{wapp_import}}}

\item{interval}{a string with the possible expressions 'hour', 'quarter', 'day', 'month', 'year' and 'weekday'}
}
\value{
new column with transformed time intervals
}
\description{
Conversion of time data in 'moment' column to time intervals
}

