% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{wapp_stat_author}
\alias{wapp_stat_author}
\title{Group an imported chat history by author}
\usage{
wapp_stat_author(chat)
}
\arguments{
\item{chat}{a WhatsApp chat history that was imported using \code{\link{wapp_import}}}
}
\value{
Grouped dataframe with informations which author writes how many posts and what is proportion of total posts
}
\description{
Group an imported chat history by author
}

