% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{wapp_stat_date}
\alias{wapp_stat_date}
\title{Aggregegates Date}
\usage{
wapp_stat_date(chat, n)
}
\arguments{
\item{chat}{Chat history that was imported with function wapp_import}

\item{n}{Top days}
}
\value{
Grouped Dataframe with Informations about the Top - n Dates with Postings
}
\description{
Aggregegates Date
}

