% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{wapp_stat_firstpost}
\alias{wapp_stat_firstpost}
\title{Who posts the first post on an day}
\usage{
wapp_stat_firstpost(chat)
}
\arguments{
\item{chat}{Chat history that was imported with function wapp_import}
}
\value{
Grouped Dataframe with Informations who posts the first post
}
\description{
Who posts the first post on an day
}

