% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{wapp_stat_time}
\alias{wapp_stat_time}
\title{Group an imported chat history by time interval}
\usage{
wapp_stat_time(chat, interval)
}
\arguments{
\item{chat}{a WhatsApp chat history that was imported using \code{\link{wapp_import}}}

\item{interval}{interval a string with the possible expressions 'hour', 'quarter', 'day', 'month', 'year' and 'weekday'}
}
\value{
Grouped dataframe with informations on which 'n' days the many posts were posted towhat is proportion of total posts
}
\description{
Group an imported chat history by time interval
}

