% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_texts.R
\name{wapp_stat_word}
\alias{wapp_stat_word}
\title{Group an imported chat history by a specific word an author has used}
\usage{
wapp_stat_word(chat, word)
}
\arguments{
\item{chat}{a WhatsApp chat history that was imported using \code{\link{wapp_import}}}

\item{word}{a specific word in the imported chat history}
}
\value{
Grouped Dataframe with informations which authors has used the word how many times
}
\description{
Group an imported chat history by a specific word an author has used
}

