% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhatsAppR.R
\name{WhatsAppR}
\alias{WhatsAppR}
\title{Visualising WhatsApp conversation data}
\usage{
WhatsAppR(file, save = FALSE, OS = "android", plot = TRUE,
  wordcloud = TRUE, CheckForErrors = FALSE)
}
\arguments{
\item{file}{the directory of the WhatsApp .txt file}

\item{save}{a logcal indicating if the output plots should be saved to the current directory}

\item{OS}{indicating if the operating system is "android" or "iOS"}

\item{plot}{indicating if plots should be created and return in output}

\item{wordcloud}{a logical indicating if a wordcloud should be plotted or not}

\item{CheckForErrors}{will check for errors in reading of input data, mostly due to quoted text}
}
\value{
Returns a named list of output plots
}
\description{
Visualising WhatsApp conversation data
}
\author{
Finlay Campbell <f.campbell15@imperial.ac.uk>
}

